Fix client loading
