+++ insights-ui/src/utils/score-utils.ts
@@ -1,0 +1,34 @@
+/**
+ * Utility functions for handling score display and styling
+ */
+
+/**
+ * Get color classes and label for a score
+ * @param score - The score value (out of 25)
+ * @returns Object containing text color class, background color class, and score label
+ */
+export function getScoreColorClasses(score: number) {
+  let textColorClass = 'text-green-500'; // Default: Green for 20+
+  let bgColorClass = 'bg-green-500'; // Default: Green for 20+
+  let scoreLabel = 'Excellent';
+
+  if (score < 8) {
+    textColorClass = 'text-red-500'; // Red for < 8
+    bgColorClass = 'bg-red-500'; // Red for < 8
+    scoreLabel = 'Poor';
+  } else if (score < 14) {
+    textColorClass = 'text-orange-500'; // Orange for 8-13
+    bgColorClass = 'bg-orange-500'; // Orange for 8-13
+    scoreLabel = 'Fair';
+  } else if (score < 20) {
+    textColorClass = 'text-yellow-500'; // Yellow for 14-19
+    bgColorClass = 'bg-yellow-500'; // Yellow for 14-19
+    scoreLabel = 'Good';
+  }
+
+  return {
+    textColorClass,
+    bgColorClass,
+    scoreLabel,
+  };
+}
